+++ docs/root/intro/version_history.rst
@@ -9,6 +9,7 @@ Version history
 * redis: added :ref:`prefix routing <envoy_api_field_config.filter.network.redis_proxy.v2.RedisProxy.prefix_routes>` to enable routing commands based on their key's prefix to different upstream.
 * redis: add support for zpopmax and zpopmin commands.
 * upstream: added :ref:`upstream_cx_pool_overflow <config_cluster_manager_cluster_stats>` for the connection pool circuit breaker.
+* access log: added RESPONSE_CODE_DETAILS.
 
 1.10.0 (Apr 5, 2019)
 ====================